#63 Improved application loading process
