+++ src/TripPlanner.App/ClientApp/src/App.tsx
@@ -21,10 +21,7 @@ import './Styles/WelcomePage.css';
 import './Styles/PlanDetailsForm.css';
 import './Styles/PlansPage.css';
 
-import React, { useEffect, useState } from 'react';
-import { Route, Routes } from 'react-router-dom';
-import { UserManager } from 'oidc-client';
-import { useSetRecoilState } from 'recoil';
+import React, { Route, Routes } from 'react-router-dom';
 import LocationActionLoadingIndicator from './components/LocationActionLoadingIndicator';
 import ModalContainer from './components/modals/ModalContainer';
 
@@ -38,117 +35,71 @@ import PlanDetailsPage from './pages/PlanDetailsPage';
 
 import PageLayout from './pages/PageLayout';
 import PrivateRoute from './components/PrivateRoute';
-import useAppSettingsService from './Services/AppSettingsService';
-import useUserManagerConfigBuilder from './Common/UserManagerConfigBuilder';
 import ErrorPage from './pages/ErrorPage';
-import { userManagerState } from './State/UserState';
 import RouteTypes from './Common/RouteTypes';
-import useLoggerService from './Services/Diagnostics/LoggerService';
-import useUserService from './Services/UserService';
 
 function App() {
-  const appSettingsService = useAppSettingsService();
-  const [isAppLoaded, setIsAppLoaded] = useState(false);
-  const setUserManager = useSetRecoilState(userManagerState);
-  const configBuilder = useUserManagerConfigBuilder();
-  const userService = useUserService();
-  const log = useLoggerService('AppInit');
-
-  useEffect(() => {
-    async function init() {
-      log.info('Start init');
-      log.info('Get app settings');
-      const settings = await appSettingsService.init();
-
-      log.info('Setup user manager');
-      const mng = new UserManager(configBuilder.build(settings));
-      setUserManager(mng);
-
-      log.info('Setup user');
-      userService.initialize(mng)
-        .then(() => {
-          log.debug('User loaded');
-        })
-        .catch(() => {
-          log.debug('No user');
-        })
-        .finally(() => {
-          log.info('End init');
-          setIsAppLoaded(true);
-        });
-    }
-
-    init();
-  }, []);
-
   return (
-    <>
-      {isAppLoaded
-        ? (
-          <div className="App">
-            <LocationActionLoadingIndicator />
-            <ModalContainer />
-            <PageLayout>
-              <Routes>
-                <Route
-                  path={RouteTypes.root}
-                  element={(
-                    <PrivateRoute>
-                      <DashboardPage />
-                    </PrivateRoute>
-                  )}
-                />
-                <Route
-                  path={RouteTypes.plans}
-                  element={(
-                    <PrivateRoute>
-                      <PlansPage />
-                    </PrivateRoute>
-                  )}
-                />
-                <Route
-                  path={RouteTypes.plan}
-                  element={(
-                    <PrivateRoute>
-                      <LocationsPage />
-                    </PrivateRoute>
-                  )}
-                />
-                <Route
-                  path={RouteTypes.callback}
-                  element={<CallbackPage />}
-                />
-                <Route
-                  path={RouteTypes.welcome}
-                  element={<WelcomePage />}
-                />
-                <Route
-                  path={RouteTypes.labPlans}
-                  element={(
-                    <PrivateRoute>
-                      <LabPlansPage />
-                    </PrivateRoute>
-                  )}
-                />
-                <Route
-                  path={RouteTypes.labPlan}
-                  element={(
-                    <PrivateRoute>
-                      <PlanDetailsPage />
-                    </PrivateRoute>
-                  )}
-                />
-                <Route
-                  path={RouteTypes.error}
-                  element={<ErrorPage />}
-                />
-              </Routes>
-            </PageLayout>
-          </div>
-        )
-      // TODO add proper loader here instead of this placeholder
-        : 'LOADING'}
-    </>
+    <div className="App">
+      <LocationActionLoadingIndicator />
+      <ModalContainer />
+      <PageLayout>
+        <Routes>
+          <Route
+            path={RouteTypes.root}
+            element={(
+              <PrivateRoute>
+                <DashboardPage />
+              </PrivateRoute>
+            )}
+          />
+          <Route
+            path={RouteTypes.plans}
+            element={(
+              <PrivateRoute>
+                <PlansPage />
+              </PrivateRoute>
+            )}
+          />
+          <Route
+            path={RouteTypes.plan}
+            element={(
+              <PrivateRoute>
+                <LocationsPage />
+              </PrivateRoute>
+            )}
+          />
+          <Route
+            path={RouteTypes.callback}
+            element={<CallbackPage />}
+          />
+          <Route
+            path={RouteTypes.welcome}
+            element={<WelcomePage />}
+          />
+          <Route
+            path={RouteTypes.labPlans}
+            element={(
+              <PrivateRoute>
+                <LabPlansPage />
+              </PrivateRoute>
+            )}
+          />
+          <Route
+            path={RouteTypes.labPlan}
+            element={(
+              <PrivateRoute>
+                <PlanDetailsPage />
+              </PrivateRoute>
+            )}
+          />
+          <Route
+            path={RouteTypes.error}
+            element={<ErrorPage />}
+          />
+        </Routes>
+      </PageLayout>
+    </div>
   );
 }
