+++ src/TripPlanner.App/ClientApp/src/AppContent.tsx
@@ -4,36 +4,38 @@ import {
 } from '@mui/material/styles';
 import { LocalizationProvider } from '@mui/lab';
 import DateAdapter from '@mui/lab/AdapterDayjs';
-import { RecoilRoot } from 'recoil';
+import { useRecoilValue } from 'recoil';
 import ToasterNotificationsComponent from './components/ToasterNotifications';
 import DebugStateObserver from './State/DebugStateObserver';
+import { isAppLoadedState } from './State/AppState';
+import App from './App';
+import AppLoader from './AppLoader';
 
 declare module '@mui/styles/defaultTheme' {
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
   interface DefaultTheme extends Theme {}
 }
 
-type Props = {
-  children: JSX.Element
-}
-
 let theme = createTheme();
 theme = responsiveFontSizes(theme);
 
-function AppContext({ children }: Props) {
+function AppContent() {
+  const isAppLoaded = useRecoilValue(isAppLoadedState);
+
   return (
-    <RecoilRoot>
+    <>
       <DebugStateObserver />
       <ToasterNotificationsComponent />
       <StyledEngineProvider injectFirst>
         <ThemeProvider theme={theme}>
           <LocalizationProvider dateAdapter={DateAdapter}>
-            {children}
+            {isAppLoaded ? <App /> : <AppLoader />}
           </LocalizationProvider>
         </ThemeProvider>
       </StyledEngineProvider>
-    </RecoilRoot>
+    </>
+
   );
 }
 
-export default AppContext;
+export default AppContent;
